+++ mottu-api/src/main/java/com/mottu/service/ModeloService.java
@@ -22,7 +22,7 @@ public class ModeloService {
     public List<Modelo> listarTodos() {
         return modeloRepository.findAll();
     }
-    
+
     @Transactional(readOnly = true)
     public Page<Modelo> listarTodos(Pageable pageable) {
         return modeloRepository.findAll(pageable);
@@ -30,42 +30,42 @@ public class ModeloService {
 
     @Transactional(readOnly = true)
     public Optional<Modelo> buscarPorId(Long id) {
-        return modeloRepository.findById(id);
+        return modeloRepository.findById(id.intValue()); // Converte Long->Integer
     }
-    
+
     @Transactional(readOnly = true)
     public List<Modelo> buscarPorFabricante(String fabricante) {
         return modeloRepository.findByFabricante(fabricante);
     }
-    
+
     @Transactional(readOnly = true)
     public Optional<Modelo> buscarPorNomeModelo(String nomeModelo) {
         return modeloRepository.findByNomeModelo(nomeModelo);
     }
-    
+
     @Transactional(readOnly = true)
     public List<Modelo> buscarPorTipo(String tipo) {
         return modeloRepository.findByTipo(tipo);
     }
-    
+
     @Transactional(readOnly = true)
     public List<Modelo> buscarPorCilindrada(Integer cilindrada) {
         return modeloRepository.findByCilindrada(cilindrada);
     }
-    
+
     @Transactional(readOnly = true)
     public List<Modelo> buscarPorCilindradaFaixa(Integer min, Integer max) {
         return modeloRepository.findByCilindradaBetween(min, max);
     }
-    
+
     @Transactional(readOnly = true)
     public List<Modelo> buscarModelosMaisUtilizados() {
         return modeloRepository.findModelosMaisUtilizados();
     }
-    
+
     @Transactional(readOnly = true)
     public long contarMotosPorModelo(Long modeloId) {
-        return modeloRepository.countMotosByModeloId(modeloId);
+        return modeloRepository.countMotosByModeloId(modeloId.intValue()); // Converte Long->Integer
     }
 
     public Modelo salvar(Modelo modelo) {
@@ -74,28 +74,25 @@ public class ModeloService {
     }
 
     public void deletar(Long id) {
-        if (!modeloRepository.existsById(id)) {
+        if (!modeloRepository.existsById(id.intValue())) {
             throw new RuntimeException("Modelo não encontrado com ID: " + id);
         }
-        
-        // Verificar se há motos vinculadas
+
         long motosVinculadas = contarMotosPorModelo(id);
         if (motosVinculadas > 0) {
             throw new RuntimeException("Não é possível deletar o modelo. Existem " + motosVinculadas + " motos vinculadas.");
         }
-        
-        modeloRepository.deleteById(id);
+
+        modeloRepository.deleteById(id.intValue());
     }
-    
+
     private void validarModelo(Modelo modelo) {
-        // Validar nome único (apenas para novos modelos ou mudança de nome)
-        if (modelo.getIdModelo() == null || !modelo.getNomeModelo().equals(buscarPorId(modelo.getIdModelo()).map(Modelo::getNomeModelo).orElse(null))) {
+        if (modelo.getIdModelo() == null || !modelo.getNomeModelo().equals(buscarPorId(modelo.getIdModelo().longValue()).map(Modelo::getNomeModelo).orElse(null))) {
             if (buscarPorNomeModelo(modelo.getNomeModelo()).isPresent()) {
                 throw new RuntimeException("Já existe um modelo com este nome: " + modelo.getNomeModelo());
             }
         }
-        
-        // Validar cilindrada
+
         if (modelo.getCilindrada() <= 0) {
             throw new RuntimeException("Cilindrada deve ser um valor positivo");
         }
